Fix type mismatch: real(8) for rem_min/rem_max in write_alya_grid_vts
The subroutine argument was declared as plain `real` (4-byte) while the
caller passes real(8) arrays, causing a gfortran type-mismatch error at
the call site.  Changed to real(8) to match the declaration in the
main program scope.

+++ AlyaProxy/alya_all2all_time_loop.f90
@@ -550,7 +550,7 @@ contains
     implicit none
     integer(4),   intent(in) :: arank, asize, ndime, itime
     MY_MPI_COMM,  intent(in) :: PAR_COMM
-    real,         intent(in) :: rem_min(3), rem_max(3)
+    real(8),      intent(in) :: rem_min(3), rem_max(3)
     integer,      intent(in) :: rem_nx(3)
     real(8),      intent(in) :: recvbuf(:)
     real(8),      intent(in) :: recvcoord(:)
